Create Mad Libs Generator
One of the best ideas to start experimenting with your hands-on python projects for students is working on Mad Libs Generator. This is the perfect project for beginners who are just starting with software development. This project will primarily focus on strings, variables, and concatenation, this project will teach you how to manipulate user-inputted data. The program design is such that it will ask users to enter a series of inputs that will be considered as a Mad Lib. Mab lib is one of the python projects for beginners. 

The input could be anything, an adjective, a noun, a pronoun, etc. Once all the inputs are entered, the application will take the data and arrange the inputs into a story template form. Sound fun, right?